i#6660 dr$sim aliases: Add -tool to replace -simulator_type
Adds -tool as the preferred alias for the -simulator_type option for
the drmemtrace/drcachesim framework.  This is long overdue as many
analysis tools are not "simulators" and the original option was solely
to pick between cache and TLB simulators.

Updates all uses in tests and documentation.

Adds a release note.

Changes the preferred cache and TLB simulator tool names to have
_simulator suffixes.  Adds a "drcachesim" alias for the cache
simulator as well.

Manually tested the "cache" and "drcachesim" aliases.

Issue: #6660