i#5623: AARCH64 feature FEAT_PAUTH2 is now recognized
Enhance the AARCH64 feature interface so that we can correctly
identify features identifed by more than one register nibble.

Update the identification of FEATURE_PAUTH so that all six
identification nibbles are checked.

Remove the automatic assumption that if the value of the nibble in
the register is greater than the expected value then the feature is
implemented.

Issue: #5623