i#3544 RV64: Rename DR_ISA_RV64IMAFDC to DR_ISA_RV64 (#6817)
RISC-V's ISA string describes a certain combination of extensions. For
example, RV64IMAFDC is a 64-bit architecture that supports I M A F D C
extensions, but there is no available hardware with this combination.
For now, the most common one is RV64GC, where G is short for
IMAFD_Zicsr_Zifencei.

DynamoRIO only relies on the G extension to work properly and it can run
on various combinations as long as all the extensions the program
requires are supported. So the name `DR_ISA_RV64IMAFDC` is not a minimum
requirement, nor a combination we supported. This PR changes it to
`DR_ISA_RV64` to be more generic.

Issue: #3544