i#6662 public traces, part 5: SYS_futex
The public version of a trace needs to preserve the
TRACE_MARKER_TYPE_FUNC_[ID | ARG | RETVAL] markers
related to SYS_futex system calls.
We add this feature to encodings2regdeps_filter_t.
This filter still drops the TRACE_MARKER_TYPE_FUNC_
markers related to other functions that are not SYS_futex.
However, we still rely on type_filter_t to remove
the additional markers that we don't want to preserve
in the public trace.

Issue #6662