i#5623 AArch64: Improve PAUTH transparency
Previously direct/indirect branches with pointer authentication would
be mangled to strip the pointer authentication code from the address
using an xpaci instruction.

This means that code that might fail when running natively (because the
pointer authentication doesn't pass) could succeed under DynamoRIO
because the pointer was not being authenticated.

This commit changes the mangling code to use auti* instructions to
authenticate the pointer instead, and adds tests to check that it
behaves correctly for all the branch and authenticate instructions.

Issue: #5623
Fixes: #5623
Co-authored-by: Phil Ramsey <phil.ramsey@arm.com>