i#6814: Set AArch64 link reg for native sig delivery
Sets the x30 link register on AArch64 appropriately to dynamorio_sigreturn when reusing the current frame for native signal delivery.

The missing x30 value caused issues in an internal test. This doesn't affect the detach_signal test likely because it uses a longjmp to return from the signal, not a sigreturn.

Issue: #6814