i#6844: Add times_of_interest to drmemtrace scheduler
Adds a new feature times_of_interest to the drmemtrace scheduler.
This allows specifying regions of interest via timestamps, which cut
across all inputs in a workload, providing for a consistent starting
point across all threads.

The feature requires a cpu schedule file which is used to build a
mapping from timestamps to instruction ordinals.  The mapping is not
perfect due to collapsed consecutive entries and coarse-grained
timestamps so interpolation is used in between known points.

Adds a unit test.

Adds a new analyzer command-line option -skip_to_timestamp which sets
a single time-of-interest in the scheduler.  Adds a test using the
checked-in threadsig x86_64 trace.

Fixes #6844