i#2416 AArch64: Fix drdecode test
On AArch53 instrlist_insert_mov_instr_addr() uses a sequence of movz and
movk instructions to, and uses the encode_estimate pointer to determine
how many movk instructions will be needed.

The test api.drdecode was calling instrlist_insert_mov_instr_addr() with
this ilist as the encode_estimate pointer which proved to be a bad
estimate on some systems.

The system where I was testing had:
    ilist =          0x000000001a3b0380
    generated_code = 0x0000ffffb4fb0000
Each movz/movk instruction in the sequence encodes a 16-bit immediate so
3 instructions would be needed to move an address in the generated_code
buffer, but using the ilist pointer as an estimate we would only generate
2 and end up with a truncated address.

Now the test will allocate the buffer for the generated code up front so
we can use the buffer pointer as an estimate instead.

Issue: #2416