i#6451 AArch64: Fix large page support
PR #5835 inadvertently broke support for large pages on AArch64 by
introducing code that assumed a 4K page size.

The core options unit_tests were also failing on large page systems
because a lot of options need to be overridden and the test was not
accounting for extra options being set in the options string.

Issues: #1680, #6451
Fixes: #6451