i#2416 AArch64: Make stolen-reg test more reliable
The test client prints a message for any load/store that uses the
stolen register as the index, assuming only the two instructions in the
test app will be found.

I have seen in at least one case the test outputs an extra line because
it found an additional instruction that used the stolen register as the
index. Presumably this instruction came from a system library so this
problem is OS dependent.

I have made the test more robust by changing the client to look for
    mov <stolen_register>, #0
before the load/store which is present in the test app but unlikely to
appear before matching load/stores in system libraries.

Issue: #2416