i#2417 AArch64: Fix ldstex mangling
The code to mangle excusive load and store instructions emits code that
has UNPREDICTALE behaviour on some AArch64 implementations.

For a sequence like this (from the client.ldstex test):

 ldaxp  (%x0)[8byte] -> %w1 %w2
 add    %w1 $0x0001 lsl $0x00 -> %w4
 add    %w2 $0x0001 lsl $0x00 -> %w5
 stlxp  %w4 %w5 -> (%x0)[8byte] %w3

The mangling would emit compare-and-swap code that does a 1x8-byte
load instead of a 2x4-byte load:

 ldaxr  (%x0)[8byte] -> %x3
 ...
 stlxp  %w4 %w5 -> (%x0)[8byte] %w3

The store stays as a 2x4-byte access. This ldex/stex pair has the same
address and access size, but access a different number of registers.
This works on some AArch64 hardware but the behaviour is UNPREDICTABLE
so this code can't be relied upon to work on all implementations:

    "An implementation of the LoadExcl and StoreExcl instructions can
    require that, in any thread of execution, the StoreExcl instruction
    accesses the same number of registers as the preceding LoadExcl
    instruction executed in that thread. If the StoreExcl instruction
    accesses a different number of registers than the preceding
    LoadExcl instruction in the same thread of execution, behavior is
    CONSTRAINED UNPREDICTABLE. As a result, software can rely on an
    LoadExcl/StoreExcl pair to eventually succeed only if they access
    the same number of registers."

    - Arm ARM DDI 0487K.a B2.17.5

On implementations that do not support ldstex pairs with different
numbers of registers, the stlxp would always fail and the app would go
in to an infinite loop.

I have changed the mangling code to always emit matching ldex/stex in
this case, and removed a section of the client.ldstex test which dealt
with mismatched register number pairs as this will hang on some hardware
even running natively.

Issue: #2417