i#6662 record_filter: parse_string() bug fix
std::numeric_limits<>::min()/max() return 0 on enum types.
This causes the parse_string() specialization for trace_marker_type_t
and trace_type_t to not add any value to the vector used by type_filter
to remove markers from a trace.
We revert this change to the previous behavior: always add the parsed value
to the vector, even if static_cast<> might truncate it.
If we want to re-add the check that type T can hold the parsed value,
we'll need to add std::underlying_type support to trace_marker_type_t
and trace_type_t enums.

Issue #6662