i#3544 RV64: Fix our memcpy return value error (#6873)
We optimized RV64 memfuncs in PR #6800. However, there was a mistake
where the return value of memcpy was incorrectly set to ARG2 which led
to segfaults in the release build. This PR addresses this issue.
Additionally, our unit_tests did not catch this error because
`ret = our_memcpy(&i, &j, sizeof(i));` was optimized by the compiler to
`i = j; ret = &i;`. We resolved this by utilizing the `noinline` attribute.