i#6882: Set drmemtrace scheduler init state to post-skip
For inputs with regions of interest that skip over their initial
records, the drmemtrace scheduler was using the initial records for
the initial timestamp and start-unscheduled state, which is incorrect.

Solves this by applying an initial region skip at init time instead of
waiting for an input to be scheduled.  This requires shifting around
when record file output is recorded.

Adds a unit test that fails without this fix as it leaves one input
permanently unscheduled and has the two inputs in the wrong order
without using the skip timestamp.

Fixes #6882