Update squash-and-merge and commit message docs (#6882)
Updates the docs on merging a pull request to the new scheme of using
the pull request title + description. The docs were still describing the
old setup where devs had to manually edit a concatenated
squash-and-merge message.

Updates the branch removal docs as that is now automated.

Adds new recommendations to keep commit message lines to 80 chars
and to allow markdown.