i#6662: Add support for REGDEPS traces to func_view
The func_view tool does not print "call" and "return" statistics for
system calls.
We add support to do so for REGDEPS synthetic traces, since SYS_futex
is the only function that is present in these traces.
A REGDEPS trace still needs a funclist.log in the trace directory
containing: 4294967498,6,,SYS_futex == ID, #args, <unused>, name
in order for func_view to work.

Tested manually, as adding funclist.log to a record_filter-generated
REGDEPS trace in suite/test/CMakeLists.txt doesn't seem trivial and
adding complexity to it doesn't seem worth it for this change.

Issue: #6662