i#6852: Mark rdvl inline asm block as volatile
Marks the inline asm block that uses the ARM SVE rdvl instruction as volatile.

On some compiler toolchains, the rdvl was not gated by the if-condition as it should be, which led to crashes on non SVE hardware. This may be because the compiler did not properly recognize the rdvl instruction encoded as raw bytes. Marking the inline asm as volatile would prevent the compiler from doing any potentially problematic transformation.

Fixes: #6852