i#6852: Mark rdvl inline asm block as volatile (#6912)
Marks the inline asm block that uses the ARM SVE rdvl instruction as
volatile.

On some compiler toolchains, the rdvl in the generated asm was not gated
by the if-condition as it should be, which led to crashes on non SVE
hardware. This may be the compiler simply deciding that it can reorder
the rdvl because it considered it as side-effect-free. Marking the
inline asm as volatile would prevent the compiler from doing any
potentially problematic transformations or reorderings.

Also tried changing the inline asm to use the mnemonic form of the rdvl
and used -march=armv8-a+sve to see if the compiler decides against
reordering, but still observed the crash on the same compiler toolchain.

Generated asm by the compiler toolchain, 
with and without volatile:
https://gist.github.com/abhinav92003/0b7baf90ac8a5e7ba44aedacf27a592e

Fixes: #6852