i#5505: Add option to skip kcore dump
Adds a new drmemtrace option that allows the user to skip creating the /proc/kcore and /proc/kallsyms dump during kernel tracing. This option may be useful in cases when it is not feasible to run the app with super-user permissions for tracing, and the user wants to use a pre-existing dump or a dump created in some other manner.

Modifies the kcore dump logic to include multiple pages worth of kcore after each bpf jit symbol. It does not seem uncommon for /proc/kallsyms to not have full page-level coverage of bpf jit code seen in the PT raw trace. This increase the kcore dump size only by a very small amount (~1%) to 20MB.

Relaxes raw2trace error handling to allow some recoverable errors during libipt decode. We keep counts of instructions that failed to decode, which were seen to cause a single instr PC discontinuity in the decoded syscall trace. We also keep counts of other non-recoverable errors that caused failure in decode of the system call. Overall, for a large app, we were still able to get >98% kernel trace decoded successfully. We will continue to debug and improve PT decode stability.

Issue: #5505