i#6685 core-shard: Add only_shards drmemtrace filter
Adds a new filter by shard ordinal to the drmemtrace scheduler and
analyzers, and adds multi-thread filtering on the command line.

Adds input_workload_t.only_shards to filter scheduler inputs by
ordinal, which is mostly useful for core-sharded-on-disk traces.

Adds new CLI options -only_shards, which takes a comma-separated list
of ordinals, and -only_threads, which takes a comma-separated list of
tids.

Adds sorting of file names when reading inputs from a directory so
that ordinals are reliable.  Adds leading 0's to record_filter's
output name so this lexicographic sort keeps cores in order.

Adds error checking that any tid or ordinal filter is actually present
in the inputs.

Adds an internal struct input_read_info_t to implement the new
filtering inside the scheduler.

Adds scheduler unit tests of the new filters.

Adds an end-to-end test of -only_shards on a core-sharded-on-disk trace.

Also tested end-to-end manually with invalid tids, invalid ordinals,
valid tids for core-sharded-on-disk, and valid ordinals for
core-sharded-on-disk.

Issue: #6685