i#6662 public traces, part8: v2p textproto reader
Implementation of a simple reader for virtual to
physical address mapping in textproto format.
The reader parses blocks in the following form:
address_mapping {
    virtual_address: 0x123
    physical_address: 0x3
}
present in the provided textproto file to create a
map in memory from virtual to physical address.

Adds unit tests and a sample v2p textproto file
to check the correctness of the generated map.

Issue #6662