i#5505 kernel trace: Add option to skip kcore dump
Adds a new drmemtrace option that allows the user to skip creating the /proc/kcore and /proc/kallsyms dump during kernel tracing. This option may be useful in cases when it is not feasible to run the app with super-user permissions for tracing, and the user wants to use a pre-existing dump or a dump created in some other manner.

Adds a test that only collects a kernel PT trace but does not dump kcore. This test is run without superuser permissions to show that superuser permissions are not required for other parts of the kernel tracing workflow.

Issue: #5505