i#5505 kernel tracing: Add option to skip kcore dump (#6932)
Adds a new drmemtrace option that allows the user to skip creating the
/proc/kcore and /proc/kallsyms dump during kernel tracing. This option
may be useful in cases when it is not feasible to run the app with
super-user permissions for tracing, and the user wants to use a
pre-existing dump or a dump created in some other manner.

Adds a test that only collects a kernel PT trace but does not dump
kcore. This test is run without superuser permissions to show that
superuser permissions are not required for other parts of the kernel
tracing workflow. Though note that perf_event_paranoid still needs to be
set to a favorable value before the test is run, which does require
superuser. This test is removed from the regular CI test suite by
appending _SUDO, to avoid confusion caused by test failure when
perf_event_paranoid is not set appropriately which may be the case when
kernel tracing code is not being intentionally tested. See comment in
CMake file.

Verified that the PT related tests continue to work on a system that
supports the Intel-PT hardware feature:

```
The following tests passed:
	code_api|tool.drcacheoff.kernel.simple_SUDO
	code_api|tool.drcacheoff.kernel.opcode-mix_SUDO
	code_api|tool.drcacheoff.kernel.syscall-mix_SUDO
	code_api|tool.drcacheoff.kernel.invariant-checker_SUDO

The following tests passed:
	code_api|client.drpttracer_SUDO-test
```

Issue: #5505