Work around the #6942 crash by always setting the disasm syntax to DR style for REGDEPS traces.
This change actually sets the disasm syntax to AT&T by default if no
module path is passed in; which is what it is supposed to do: but it
was not doing that and this breaks 3 tests comparing DR-style output.
We put in a quick fix to request DR style for those tests.

Issue: #6942