i#6940: Do not always load binaries in view tool (#6941)
The view tool was blindly loading binaries even for traces that have
encodings.  This leads to fatal errors when binaries have changed, even
when the change has no impact on viewing a trace.  We fix that here by
reading the filtype at init time.

Work around the #6942 crash by always setting the disasm syntax to DR
style for REGDEPS traces.
    
This change actually sets the disasm syntax to AT&T by default if no
module path is passed in; which is what it is supposed to do: but it was not
doing that and this breaks 3 tests comparing DR-style output. We put in a
quick fix to request DR style for those tests.

Tested locally where the view tool asserts without this fix.

Issue: #6940, #6942
Fixes #6940