i#6945: Reduce default block time
Reduces the scheduler and drmemtrace launcher default values for
block_time_scale down to 10 and block_time_max down to 2.5s.  This
improves the scheduler behavior for small traces under fast analyzers.
It seems better to err on the side of faster and let more heavyweight
simulations tune the block times for more idle time; otherwise we can
end up with local runs and especially new users trying things out and
seeing the tool seem to just sit there doing nothing.

This reduces the threadsig core-sharded time from a minute and a half
down to 10 seconds in local runs (see #6945 for command lines); there
is still some idle time in there so it seems a reasonable compromise.

Fixes #6945