i#6959 all-unsched: Revert only-unscheduled fix
Reverts a fix for a bug in the scheduler where it let a thread going
unscheduled continue running if there are no other non-running-now
scheduleable inputs.  This triggered too-frequent all-unscheduled
cases and the current timeout for those is too high, causing tail
delays.  We'll re-instate the fix once we add an early exit feature
for that scenario.

Issue: #6959