i#6662 public traces, part 9: TLB simulator with v2p
We want to provide a virtual to physical mapping for
the release of public traces to be used with our TLB
simulator.
We enable the TLB simulator to operate on the physical
addresses provided by a v2p.textproto file, which is
read by the previously added v2p_reader.

Issue #6662