i#4501: Undo some changes to loader_shared.c made by 9293e7ae7.
The purpose of 9293e7ae7 was to add TLS support for Windows,
but it broke the use of multiple clients. This patch brings
back the previous code for non-Windows, so on Unix all the
loading happens later, in loader_init_epilogue().

Also add a test of two clients.

Unfortunately this patch undoes the unification of the Unix
and Windows loaders that happened with 9293e7ae7. It would
be good to find a way of unifying the Unix and Windows loaders
that does not break the use of multiple clients (#6963).

Change-Id: I6bd63de90a1d9eebe50b04ac6117cc3522ae7df7