i#6938 sched migrate: Add lock contention stats
Adds lock contention counters for mutex_dbg_owned::lock() in
non-NDEBUG builds.  While it's not there for release builds, the
results in non-release should still be indicative of general lock
behavior.

Prints the stats for sched_lock_ with the other scheduler stats.

Sample results on schedule_stats on a threadsig trace show a lot of
contention even with only 3 cores:

```
$ clients/bin64/drmemtrace_launcher -indir ../build_x64_dbg_tests/drmemtrace.threadsig.5* -core_sharded -cores 3 -tool schedule_stats -verbose 1
[scheduler] Schedule lock acquired     :   2196602
[scheduler] Schedule lock contended    :    257580
```

Issue: #6938