i#6981: Use consistent version for upload-artifact and download-artifact (#6983)
Changes the version used for upload-artifact and download-artifact
actions to the current one recommended at
https://github.com/actions/upload-artifact and
https://github.com/actions/download-artifact.

For these two actions, "v4" points to the latest available v4.x.x
release, as evidenced by the commit hash for the versions at
https://github.com/actions/download-artifact/tags and
https://github.com/actions/upload-artifact/tags. So we would still be
using the latest version as suggested by the security advisory for
download-artifact (https://github.com/advisories/GHSA-cxww-7g56-2vh6).

PR #6964 bumped the download-artifact version but not the
upload-artifact which caused a mismatch and the #6981 issue.

Test run worked fine:
https://github.com/DynamoRIO/dynamorio/actions/runs/10835746272

 Fixes: #6981