i#6971: Use instr count instead of wallclock for simulated time
When using the drmemtrace scheduler in an analyzer or other tool that
does not track simulated time, the scheduler used to use wall-clock
time.  Here we change that to use the instruction count plus a scaled
idle count.  An idle counter is added and a new scale option
scheduler_options_t.time_units_per_idle (and CLI
-sched_time_units_per_idle) defaulting to 5.

The time_units_per_us and sched_time_units_per_us defaults are set to
1000, reflecting a gHz machine with IPC=0.5

Using counters provides a more reproducible result across different
runs and machines.

Adds a test of the new option.

The default values of the options were tested on a large trace and
found to produce a representative level of idle time during the main
execution (and the whole run when combined with the forthcoming
exit-early feature for #6959).

This means that the clock going backward problem (#6966) is no longer
seen in default runs.  The analyzer still supports wall-clock with the
-sched_time option so a check to avoid underflow is added.

Fixes #6971
Fixes #6966