Remove time_units_per_idle and instead raise sched_exit_if_fraction_left to 0.25
The conclusion is that the old time_units_per_us=100 for wall clock
was too low; to match it with counts, either the time_units_per_idle
scaling is needed or just dropping time_units_per_us from 1000 to 500
but that seems unrealistic.  Instead we can get the results we want
from our large traces by exiting earlier, since most of the unwanted
idle is still in seemingly unrepresentative regions at the end.