i#7031: invalidate CPU marker filter
In some cases we don't want to expose the "as traced" CPU schedule
in an offline trace because it might not be representative of the
native execution of the traced program.

To do so we implement `invalidate_cpu_filter_t`, a new filter that
is used with `record_filter` as follows:
```
drrun -t drmemtrace -tool record_filter -filter_invalidate_cpu -indir path/to/input/trace -outdir path/to/output/trace
```
This filter set the value of TRACE_MARKER_TYPE_CPU_ID markers to
(uintptr_t)-1, which representes an unknown CPU.

We add a unit test `test_invalidate_cpu_filter()` and an end-to-end
test `code_api|tool.record_filter_invalidate_cpu` which invokes the
`invariant_checker` and `view` tool on the filtered trace.

Fixes #7031