i#7028: Modify core-sharded cpu ids and timestamps
In core-sharded dynamically-scheduled mode, the as-traced cpu ids and
timestamps are interleaved in a way that makes them confusing and not
useful.  We change the scheduler here to replace each cpu id value
with the shard id, and replace the timestamp with a synthetic value
based on the instruction and idle counts.

Adds a test.

Fixes #7028