i#7023: Switch drmemtrace replay of idles to use record count
The drmemtrace replay of idles was using wall-clock time to reproduce
the duration.  A new idle replay type is added here (and made the
default) which uses the idle record count instead, leaving support for the old
format under the old type.

Tested the new idle type on the existing record-replay tests inside
scheduler_unit_tests and in core_sharded_test.

Support for the old format is considered best-effort as its use of
wall-clock time makes it not able to reproduce precise timing in any
case.  Tested manually by generating an old-style record and ensuring
the new code handled it:

  $ clients/bin64/drmemtrace_launcher -indir ../src/clients/drcachesim/tests/drmemtrace.threadsig.x64.tracedir -core_sharded -cores 4 -replay_file ../build_x64_rel_tests/legacy_record.zip

Fixes #7023