i#6938 sched migrate: Enforce migration threshold at the start
Previously, a never-executed input could be moved to another output at
any time, yet was still counted as a "migration".  We change that here
to consider a never-executed input to have executed at the initial
simulation time seen on an output, so it will not be migrated until
that threshold is met.  An exception is the very first rebalance at
init time for the initial allocation of inputs to outputs when inputs
can be freely moved; this does not count as a migration.

Adds a unit test.

Issue: #6938