i#7044 hang: Fix deadlock regression on drmemtrace replay (#7049)
Fixes a deadlock on replay introduced by PR #6985. If a recorded
schedule has back-to-back entries with the same input (with a gap where
another thread runs that input in between of course) and it has to wait
for that other thread for its 2nd stint with that input, it will
deadlock as it holds the input's lock while it tries to get the lock in
set_cur_input(). That 2nd lock acquisition is what was added by PR
#6985: that's the source of the regression.

Tested on a 12-thread threadsig trace which hangs without the fix and
succeeds with the fix. It is too large to turn into a convenient
regression test.

Issue: #7044