i#7044 hang: Sleep on wait in parallel drmemtrace analyzer
Adds a 1ms sleep on a STATUS_WAIT in the drmemtrace analyzer when
operating in parallel.  This avoids costly spinning in replay mode.
Adding targeted condition variables inside the scheduler doesn't end
up performing any better and is complicated by the scheduler not
knowing the threading model of the user, so this is a safer and
simpler solution despite not waiting on an event.

Fixes #7044