i#6938 sched migrate: Add observed migrations to schedule stats
While the schedule_stats tool already reports the migration count from
the scheduler, that is only non-zero for a dynamic schedule, resulting
in 0 for core-sharded-on-disk traces.  We add "observed migrations"
where schedule_stats looks for migrations on its own.  This requires a
global lock on each context switch, but experimental results show that
this does not seem to cause noticeable overhead.

Adds some testing.

Issue: #6938